qt-cpp: Refactor test runner to extract common helpers
Extract shared code from runTest.mts and runTest.build.mts into a new
runTestHelper.mts module to eliminate duplication and avoid bundling
vscode dependency in Node.js test runner context.

- Create runTestHelper.mts with Node.js-compatible helper functions
- Move getCliArg, getQtRoot, verifyQtCoreVsix, setupVSCodeSettings,
  installRequiredExtensions, and setupTestInfrastructure to new module
- Update both runTest.mts and runTest.build.mts to use extracted helpers
- Keep helper.mts focused on VS Code extension host test utilities
